+++ src/main/java/com/springBoard/SpringBoardApplication.java
@@ -1,0 +1,13 @@
+package com.springBoard;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpringBoardApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SpringBoardApplication.class, args);
+	}
+
+}
