+++ src/main/java/com/springBoard/board/model/Board.java
@@ -1,0 +1,4 @@
+package com.springBoard.board.model;
+
+public class Board {
+}
