+++ src/main/java/com/springBoard/board/repository/BoardRepository.java
@@ -1,0 +1,4 @@
+package com.springBoard.board.repository;
+
+public interface BoardRepository {
+}
