+++ src/main/java/com/springBoard/board/service/BoardService.java
@@ -1,0 +1,4 @@
+package com.springBoard.board.service;
+
+public interface BoardService {
+}
