+++ src/main/java/com/springBoard/comment/model/Comment.java
@@ -1,0 +1,4 @@
+package com.springBoard.comment.model;
+
+public class Comment {
+}
