+++ src/main/java/com/springBoard/comment/repository/CommentRepository.java
@@ -1,0 +1,4 @@
+package com.springBoard.comment.repository;
+
+public interface CommentRepository {
+}
