+++ src/main/java/com/springBoard/comment/service/CommentService.java
@@ -1,0 +1,4 @@
+package com.springBoard.comment.service;
+
+public interface CommentService {
+}
