+++ src/main/java/com/springBoard/config/WebMvcConfig.java
@@ -1,0 +1,4 @@
+package com.springBoard.config;
+
+public class WebMvcConfig {
+}
