+++ src/main/java/com/springBoard/controller/PostController.java
@@ -1,0 +1,4 @@
+package com.springBoard.controller;
+
+public class PostController {
+}
