+++ src/main/java/com/springBoard/exception/AccessDeniedException.java
@@ -1,0 +1,4 @@
+package com.springBoard.exception;
+
+public class AccessDeniedException {
+}
