+++ src/main/java/com/springBoard/payload/ApiResponse.java
@@ -1,0 +1,4 @@
+package com.springBoard.payload;
+
+public class ApiResponse {
+}
