+++ src/main/java/com/springBoard/post/model/Post.java
@@ -1,0 +1,4 @@
+package com.springBoard.post.model;
+
+public class Post {
+}
