+++ src/main/java/com/springBoard/post/repository/PostRepository.java
@@ -1,0 +1,4 @@
+package com.springBoard.post.repository;
+
+public interface PostRepository {
+}
