+++ src/main/java/com/springBoard/post/service/PostService.java
@@ -1,0 +1,4 @@
+package com.springBoard.post.service;
+
+public interface PostService {
+}
