+++ src/main/java/com/springBoard/user/model/User.java
@@ -1,0 +1,4 @@
+package com.springBoard.user.model;
+
+public class User {
+}
