+++ src/main/java/com/springBoard/user/repository/UserRepository.java
@@ -1,0 +1,4 @@
+package com.springBoard.user.repository;
+
+public interface UserRepository {
+}
