+++ src/main/java/com/springBoard/user/service/UserService.java
@@ -1,0 +1,4 @@
+package com.springBoard.user.service;
+
+public interface UserService {
+}
