--- src/main/java/sungminyun/springBoard/SpringBoardApplication.java
@@ -1,13 +1,0 @@
-package sungminyun.springBoard;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SpringBoardApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SpringBoardApplication.class, args);
-	}
-
-}
