--- src/main/java/sungminyun/springBoard/config/WebMvcConfig.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.config;
-
-public class WebMvcConfig {
-}
