--- src/main/java/sungminyun/springBoard/controller/BoardController.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.controller;
-
-public class BoardController {
-}
