--- src/main/java/sungminyun/springBoard/exception/AccessDeniedException.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.exception;
-
-public class AccessDeniedException {
-}
