--- src/main/java/sungminyun/springBoard/model/Board.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.model;
-
-public class Board {
-}
