--- src/main/java/sungminyun/springBoard/model/Comment.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.model;
-
-public class Comment {
-}
