--- src/main/java/sungminyun/springBoard/model/User.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.model;
-
-public class User {
-}
