--- src/main/java/sungminyun/springBoard/payload/ApiResponse.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.payload;
-
-public class ApiResponse {
-}
