--- src/main/java/sungminyun/springBoard/repository/BoardRepository.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.repository;
-
-public interface BoardRepository {
-}
