--- src/main/java/sungminyun/springBoard/repository/CommentRepository.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.repository;
-
-public interface CommentRepository {
-}
