--- src/main/java/sungminyun/springBoard/repository/PostRepository.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.repository;
-
-public interface PostRepository {
-}
