--- src/main/java/sungminyun/springBoard/service/BoardService.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.service;
-
-public interface BoardService {
-}
