--- src/main/java/sungminyun/springBoard/service/CommentService.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.service;
-
-public interface CommentService {
-}
