--- src/main/java/sungminyun/springBoard/service/PostService.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.service;
-
-public interface PostService {
-}
