+++ src/test/java/com/springBoard/SpringBoardApplicationTests.java
@@ -1,0 +1,13 @@
+package com.springBoard;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SpringBoardApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
