--- src/test/java/sungminyun/springBoard/SpringBoardApplicationTests.java
@@ -1,13 +1,0 @@
-package sungminyun.springBoard;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringBoardApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
